# Jorje.TheWorld.Api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "v1",
    "title": "Trips API"
  },
  "paths": {
...
    "/api/stops": {
      "get": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "PageSize",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "SearchQuery",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "OrderBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Fields",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Accept",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsPost",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "stopInput",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/StopForCreationDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "options": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsOptions",
        "consumes": [],
        "produces": [],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "head": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsHead",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "PageNumber",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "PageSize",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "SearchQuery",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "OrderBy",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Fields",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "Accept",
            "in": "header",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/stops/{id}": {
      "get": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsByIdGet",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "fields",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsByIdPut",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "stopInput",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/StopForUpdateDTO"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsByIdDelete",
        "consumes": [],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": [
          "Stops"
        ],
        "operationId": "ApiStopsByIdPatch",
        "consumes": [
          "application/json",
          "text/json",
          "application/json-patch+json",
          "application/xml",
          "text/xml"
        ],
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "patchDoc",
            "in": "body",
            "required": false,
            "schema": {
              "$ref": "#/definitions/JsonPatchDocument[StopForUpdateDTO]"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "StopForCreationDTO": {
      "required": [
        "name",
        "description"
      ],
      "type": "object",
      "properties": {
        "name": {
          "maxLength": 100,
          "type": "string"
        },
        "description": {
          "maxLength": 500,
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "StopForUpdateDTO": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "latitude": {
          "format": "double",
          "type": "number"
        },
        "longitude": {
          "format": "double",
          "type": "number"
        }
      }
    },
    "JsonPatchDocument[StopForUpdateDTO]": {
      "type": "object",
      "properties": {
        "operations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Operation[StopForUpdateDTO]"
          },
          "readOnly": true
        }
      }
    },
    "Operation[StopForUpdateDTO]": {
      "type": "object",
      "properties": {
        "value": {
          "type": "object"
        },
        "path": {
          "type": "string"
        },
        "op": {
          "type": "string"
        },
        "from": {
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}